+++ src/common.c
@@ -1,3 +1,4 @@
+/* SPDX-License-Identifier: MPL-2.0 */
 #include "internal.h"
 
 #include <stdio.h>